[test] fix msvc build error

+++ tests/017-serialization/main.cpp
@@ -52,7 +52,7 @@ vcl::Color randomColor()
     // generate random color using std::mt19937
 
     std::mt19937 gen;
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
+    std::uniform_int_distribution<uint> dist(0, 255);
     return vcl::Color(dist(gen), dist(gen), dist(gen), dist(gen));
 }
